cleaning up the code using -Werror I cannot finish it in Debug mode becuase of my macros though

+++ ParU/Source/paru_intersection.cpp
@@ -41,7 +41,7 @@ int paru_intersection ( Int e, paru_Element **elementList,
     PRLEVEL (p, ("%% newColSet.size = %ld\n",stl_newColSet.size() ));
     PRLEVEL (p, ("%% nEl = %ld\n",nEl));
     std::set<Int>::iterator it;
-    if (C*stl_newColSet.size() < nEl-el->lac )
+    if ( (Int) (C*stl_newColSet.size()) < nEl-el->lac )
         // if size el >> stl_newColSet 
         //   binary search each of elements in stl_newColSet in el
         //   log(nEl)*stl_newColSet.size()
@@ -61,7 +61,7 @@ int paru_intersection ( Int e, paru_Element **elementList,
             };
         }
     }
-    else if (stl_newColSet.size() > C*(nEl-el->lac) )
+    else if ( (Int) stl_newColSet.size() > C*(nEl-el->lac) )
     {   //  else if stl_newColSet >> el
         PRLEVEL (p, ("%%el << stl_newColSet\n"));
         //      binary search each of elements in el in stl_newColSet
